+++ lms/static/sass/bootstrap/_components.scss
@@ -14,6 +14,10 @@
     list-style: none;
     padding: 0;
     margin: 0;
+
+    li {
+      margin: 5px 0;
+    }
   }
 
 }
