+++ lms/static/sass/elements/_banners.scss
@@ -63,6 +63,10 @@ $full-width-banner-margin: 20px;
     list-style: none;
     padding: 0;
     margin: 0;
+
+    li {
+      margin: 5px 0;
+    }
   }
 
   }
