+++ openedx/features/content_type_gating/block_transformers.py
@@ -7,7 +7,7 @@ from django.conf import settings
 from openedx.core.djangoapps.content.block_structure.transformer import (
     BlockStructureTransformer,
 )
-from openedx.features.content_type_gating.partitions import CONTENT_GATING_PARTITION_ID
+from openedx.features.content_type_gating.helpers import CONTENT_GATING_PARTITION_ID
 from openedx.features.content_type_gating.models import ContentTypeGatingConfig
